Move rpu data
